Move to_csv to a concern
