+++ app/models/activity_log_item.rb
@@ -4,20 +4,11 @@ require "csv"
 #
 # @see RecordAction
 class ActivityLogItem < ApplicationRecord
+  include Csvable
+
   self.table_name = "activity_log"
 
   default_scope { order(:created_at) }
 
   validates :user_id, :journey_id, :action, presence: true
-
-  # @return [String]
-  def self.to_csv
-    CSV.generate(headers: true) do |csv|
-      csv << column_names
-
-      find_each do |record|
-        csv << record.attributes.values
-      end
-    end
-  end
 end
